+++ packages/evm/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [1.0.3](https://github.com/sygmaprotocol/sygma-sdk/compare/evm-v1.0.2...evm-v1.0.3) (2024-08-20)
+
+
+### Bug Fixes
+
+* Add low balance validation  ([#481](https://github.com/sygmaprotocol/sygma-sdk/issues/481)) ([db4fcaf](https://github.com/sygmaprotocol/sygma-sdk/commit/db4fcaf879e673674bd5c1cf97b86bab59c4a0ff))
+
 ## [1.0.2](https://github.com/sygmaprotocol/sygma-sdk/compare/evm-v1.0.1...evm-v1.0.2) (2024-08-14)